third_party/WebKit/LayoutTests/inspector/console/console-edit-property-value-expected.txt

CONSOLE MESSAGE: line 10: [object Object]
Tests that property values can be edited inline in the console via double click.

Node was hidden after dblclick: true
Node was hidden after dblclick: true
Node was hidden after dblclick: true
Node was hidden after dblclick: true
logToConsole()
console-edit-property-value.html:10 Object {a: 1, b: "foo", c: null, d: 2}
    a: 3
    b: "foo"
    c: Array(3)
    d: Object
    __proto__: Object
undefined
